refactor: move openingTimes
